Improve formatting, update volumes and implement build.sh
The virutalenv environment is now created in /opt/pydio-sync-venv.
This is the "final" path where the environment will be deployed to.

Changed the pydio-sync source path from /tmp to /root. It just seemed
nicer, but should be irrelevant.

build.sh now creates the virtualenv environment which can be used by the
Python2 host interpreter (like Debian/bullseye) where Python2 packages
(via pip install) can not be installed to be able to run the final
pydio-sync agent.

build.sh currently does *not* build the pydio-sync agent (only the
virtualenv environment need to run it).